04/pyproject.toml
[project]
name = "univ-oop-prac-2_04"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.11.2"
dependencies = [
    "ffmpeg-python>=0.2.0",
    "mlx-whisper>=0.4.3",
    "numpy>=2.3.4",
    "pydub>=0.25.1",
    "typed-ffmpeg>=3.7.1",
]

[tool.ruff]
line-length = 100

[tool.ruff.format]
docstring-code-format = true

# ref: https://zenn.dev/enven/articles/python-ruff-with-vscode
[tool.ruff.lint]
select = ["ALL"]
ignore = [
  "ANN204", # missing return type annotation for magic method
  "D1",     # undocumented
  "D203",   # one blank line before class
  "D213",   # multi-line summary second line
  "D400",   # first line should end with a period
  "D415",   # first line should be capitalized
  "TD001",  # invalid todo tag
  "TD002",  # missing todo author
  "TD003",  # missing todo link
  "T201",   # print found
  "ERA001", # Found commented-out code
  "PD011",  # pandas use of dot values
  "RUF001",
  "RUF002",
  "RUF003",
  "EM101",
  "EM102",
  "EM103",
  "TRY003",
]
unfixable = [
  "F841", # unused variable
]

logger-objects = ["src.library.logger.LOGGER"]

[tool.ruff.lint.pylint]
max-args = 6
